+++ lib/src/list/built_list.dart
@@ -27,7 +27,7 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   ///
   /// Rejects nulls. Rejects elements of the wrong type.
   factory BuiltList([Iterable iterable = const []]) =>
-      new BuiltList<E>.from(iterable);
+      BuiltList<E>.from(iterable);
 
   /// Instantiates with elements from an [Iterable].
   ///
@@ -42,7 +42,7 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
     if (iterable is _BuiltList && iterable.hasExactElementType(E)) {
       return iterable as BuiltList<E>;
     } else {
-      return new _BuiltList<E>.copyAndCheckTypes(iterable);
+      return _BuiltList<E>.copyAndCheckTypes(iterable);
     }
   }
 
@@ -55,28 +55,28 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
     if (iterable is _BuiltList<E> && iterable.hasExactElementType(E)) {
       return iterable;
     } else {
-      return new _BuiltList<E>.copyAndCheckForNull(iterable);
+      return _BuiltList<E>.copyAndCheckForNull(iterable);
     }
   }
 
   /// Creates a [ListBuilder], applies updates to it, and builds.
-  factory BuiltList.build(updates(ListBuilder<E> builder)) =>
-      (new ListBuilder<E>()..update(updates)).build();
+  factory BuiltList.build(Function(ListBuilder<E>) updates) =>
+      (ListBuilder<E>()..update(updates)).build();
 
   /// Converts to a [ListBuilder] for modification.
   ///
   /// The `BuiltList` remains immutable and can continue to be used.
-  ListBuilder<E> toBuilder() => new ListBuilder<E>(this);
+  ListBuilder<E> toBuilder() => ListBuilder<E>(this);
 
   /// Converts to a [ListBuilder], applies updates to it, and builds.
-  BuiltList<E> rebuild(updates(ListBuilder<E> builder)) =>
+  BuiltList<E> rebuild(Function(ListBuilder<E>) updates) =>
       (toBuilder()..update(updates)).build();
 
   @override
   BuiltList<E> toBuiltList() => this;
 
   @override
-  BuiltSet<E> toBuiltSet() => new BuiltSet<E>(this);
+  BuiltSet<E> toBuiltSet() => BuiltSet<E>(this);
 
   /// Deep hashCode.
   ///
@@ -84,9 +84,7 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   /// the same order. Then, the `hashCode` is guaranteed to be the same.
   @override
   int get hashCode {
-    if (_hashCode == null) {
-      _hashCode = hashObjects(_list);
-    }
+    _hashCode ??= hashObjects(_list);
     return _hashCode;
   }
 
@@ -113,7 +111,7 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   ///
   /// Useful when producing or using APIs that need the [List] interface. This
   /// differs from [toList] where mutations are explicitly disallowed.
-  List<E> asList() => new List<E>.unmodifiable(_list);
+  List<E> asList() => List<E>.unmodifiable(_list);
 
   // List.
 
@@ -122,7 +120,7 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
 
   /// As [List.+].
   BuiltList<E> operator +(BuiltList<E> other) =>
-      new _BuiltList<E>.withSafeList(_list + other._list);
+      _BuiltList<E>.withSafeList(_list + other._list);
 
   /// As [List.length].
   @override
@@ -138,16 +136,16 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   int lastIndexOf(E element, [int start]) => _list.lastIndexOf(element, start);
 
   /// As [List.indexWhere].
-  int indexWhere(bool test(E element), [int start = 0]) =>
+  int indexWhere(bool Function(E) test, [int start = 0]) =>
       _list.indexWhere(test, start);
 
   /// As [List.lastIndexWhere].
-  int lastIndexWhere(bool test(E element), [int start]) =>
+  int lastIndexWhere(bool Function(E) test, [int start]) =>
       _list.lastIndexWhere(test, start);
 
   /// As [List.sublist] but returns a `BuiltList<E>`.
   BuiltList<E> sublist(int start, [int end]) =>
-      new _BuiltList<E>.withSafeList(_list.sublist(start, end));
+      _BuiltList<E>.withSafeList(_list.sublist(start, end));
 
   /// As [List.getRange].
   Iterable<E> getRange(int start, int end) => _list.getRange(start, end);
@@ -161,41 +159,41 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   Iterator<E> get iterator => _list.iterator;
 
   @override
-  Iterable<T> map<T>(T f(E e)) => _list.map(f);
+  Iterable<T> map<T>(T Function(E) f) => _list.map(f);
 
   @override
-  Iterable<E> where(bool test(E element)) => _list.where(test);
+  Iterable<E> where(bool Function(E) test) => _list.where(test);
 
   @override
   Iterable<T> whereType<T>() => _list.whereType<T>();
 
   @override
-  Iterable<T> expand<T>(Iterable<T> f(E e)) => _list.expand(f);
+  Iterable<T> expand<T>(Iterable<T> Function(E) f) => _list.expand(f);
 
   @override
   bool contains(Object element) => _list.contains(element);
 
   @override
-  void forEach(void f(E element)) => _list.forEach(f);
+  void forEach(void Function(E) f) => _list.forEach(f);
 
   @override
-  E reduce(E combine(E value, E element)) => _list.reduce(combine);
+  E reduce(E Function(E, E) combine) => _list.reduce(combine);
 
   @override
-  T fold<T>(T initialValue, T combine(T previousValue, E element)) =>
+  T fold<T>(T initialValue, T Function(T, E) combine) =>
       _list.fold(initialValue, combine);
 
   @override
   Iterable<E> followedBy(Iterable<E> other) => _list.followedBy(other);
 
   @override
-  bool every(bool test(E element)) => _list.every(test);
+  bool every(bool Function(E) test) => _list.every(test);
 
   @override
   String join([String separator = '']) => _list.join(separator);
 
   @override
-  bool any(bool test(E element)) => _list.any(test);
+  bool any(bool Function(E) test) => _list.any(test);
 
   /// As [Iterable.toList].
   ///
@@ -206,8 +204,7 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   /// This allows efficient use of APIs that ask for a mutable collection
   /// but don't actually mutate it.
   @override
-  List<E> toList({bool growable = true}) =>
-      new CopyOnWriteList<E>(_list, growable);
+  List<E> toList({bool growable = true}) => CopyOnWriteList<E>(_list, growable);
 
   @override
   Set<E> toSet() => _list.toSet();
@@ -222,13 +219,13 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   Iterable<E> take(int n) => _list.take(n);
 
   @override
-  Iterable<E> takeWhile(bool test(E value)) => _list.takeWhile(test);
+  Iterable<E> takeWhile(bool Function(E) test) => _list.takeWhile(test);
 
   @override
   Iterable<E> skip(int n) => _list.skip(n);
 
   @override
-  Iterable<E> skipWhile(bool test(E value)) => _list.skipWhile(test);
+  Iterable<E> skipWhile(bool Function(E) test) => _list.skipWhile(test);
 
   @override
   E get first => _list.first;
@@ -240,15 +237,15 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   E get single => _list.single;
 
   @override
-  E firstWhere(bool test(E element), {E orElse()}) =>
+  E firstWhere(bool Function(E) test, {E Function() orElse}) =>
       _list.firstWhere(test, orElse: orElse);
 
   @override
-  E lastWhere(bool test(E element), {E orElse()}) =>
+  E lastWhere(bool Function(E) test, {E Function() orElse}) =>
       _list.lastWhere(test, orElse: orElse);
 
   @override
-  E singleWhere(bool test(E element), {E orElse()}) =>
+  E singleWhere(bool Function(E) test, {E Function() orElse}) =>
       _list.singleWhere(test, orElse: orElse);
 
   @override
@@ -261,7 +258,7 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
 
   BuiltList._(this._list) {
     if (E == dynamic) {
-      throw new UnsupportedError(
+      throw UnsupportedError(
           'explicit element type required, for example "new BuiltList<int>"');
     }
   }
@@ -272,22 +269,39 @@ class _BuiltList<E> extends BuiltList<E> {
   _BuiltList.withSafeList(List<E> list) : super._(list);
 
   _BuiltList.copyAndCheckTypes([Iterable iterable = const []])
-      : super._(new List<E>.from(iterable, growable: false)) {
+      : super._(List<E>.from(iterable, growable: false)) {
     for (var element in _list) {
       if (element is! E) {
-        throw new ArgumentError('iterable contained invalid element: $element');
+        throw ArgumentError('iterable contained invalid element: $element');
       }
     }
   }
 
   _BuiltList.copyAndCheckForNull(Iterable<E> iterable)
-      : super._(new List<E>.from(iterable, growable: false)) {
+      : super._(List<E>.from(iterable, growable: false)) {
     for (var element in _list) {
       if (identical(element, null)) {
-        throw new ArgumentError('iterable contained invalid element: null');
+        throw ArgumentError('iterable contained invalid element: null');
       }
     }
   }
 
   bool hasExactElementType(Type type) => E == type;
+}
+
+/// Extensions for [BuiltList] on [List].
+extension BuiltListExtension<T> on List<T> {
+  /// Converts to a [BuiltList].
+  BuiltList<T> build() {
+    // We know a `List` is not a `BuiltList`, so we have to copy.
+    return _BuiltList<T>.copyAndCheckForNull(this);
+  }
+}
+
+/// Extensions for [BuiltList] on [Iterable].
+extension BuiltListIterableExtension<E> on Iterable<E> {
+  /// Converts to a [BuiltList].
+  ///
+  /// Just returns the [Iterable] if it is already a `BuiltList<E>`.
+  BuiltList<E> toBuiltList() => BuiltList<E>.of(this);
 }