+++ lib/src/list_multimap/built_list_multimap.dart
@@ -18,7 +18,7 @@ abstract class BuiltListMultimap<K, V> {
   final Map<K, BuiltList<V>> _map;
 
   // Precomputed.
-  final BuiltList<V> _emptyList = new BuiltList<V>();
+  final BuiltList<V> _emptyList = BuiltList<V>();
 
   // Cached.
   int _hashCode;
@@ -44,26 +44,27 @@ abstract class BuiltListMultimap<K, V> {
     } else if (multimap is Map ||
         multimap is ListMultimap ||
         multimap is BuiltListMultimap) {
-      return new _BuiltListMultimap<K, V>.copyAndCheck(
+      return _BuiltListMultimap<K, V>.copyAndCheck(
           multimap.keys, (k) => multimap[k]);
     } else {
-      throw new ArgumentError(
-          'expected Map, ListMultimap or BuiltListMultimap, '
+      throw ArgumentError('expected Map, ListMultimap or BuiltListMultimap, '
           'got ${multimap.runtimeType}');
     }
   }
 
   /// Creates a [ListMultimapBuilder], applies updates to it, and builds.
-  factory BuiltListMultimap.build(updates(ListMultimapBuilder<K, V> builder)) =>
-      (new ListMultimapBuilder<K, V>()..update(updates)).build();
+  factory BuiltListMultimap.build(
+          Function(ListMultimapBuilder<K, V>) updates) =>
+      (ListMultimapBuilder<K, V>()..update(updates)).build();
 
   /// Converts to a [ListMultimapBuilder] for modification.
   ///
   /// The `BuiltListMultimap` remains immutable and can continue to be used.
-  ListMultimapBuilder<K, V> toBuilder() => new ListMultimapBuilder<K, V>(this);
+  ListMultimapBuilder<K, V> toBuilder() => ListMultimapBuilder<K, V>(this);
 
   /// Converts to a [ListMultimapBuilder], applies updates to it, and builds.
-  BuiltListMultimap<K, V> rebuild(updates(ListMultimapBuilder<K, V> builder)) =>
+  BuiltListMultimap<K, V> rebuild(
+          Function(ListMultimapBuilder<K, V>) updates) =>
       (toBuilder()..update(updates)).build();
 
   /// Converts to a [Map].
@@ -74,7 +75,7 @@ abstract class BuiltListMultimap<K, V> {
   ///
   /// This allows efficient use of APIs that ask for a mutable collection
   /// but don't actually mutate it.
-  Map<K, BuiltList<V>> toMap() => new CopyOnWriteMap<K, BuiltList<V>>(_map);
+  Map<K, BuiltList<V>> toMap() => CopyOnWriteMap<K, BuiltList<V>>(_map);
 
   /// Deep hashCode.
   ///
@@ -83,12 +84,10 @@ abstract class BuiltListMultimap<K, V> {
   /// to be the same.
   @override
   int get hashCode {
-    if (_hashCode == null) {
-      _hashCode = hashObjects(_map.keys
-          .map((key) => hash2(key.hashCode, _map[key].hashCode))
-          .toList(growable: false)
-            ..sort());
-    }
+    _hashCode ??= hashObjects(_map.keys
+        .map((key) => hash2(key.hashCode, _map[key].hashCode))
+        .toList(growable: false)
+          ..sort());
     return _hashCode;
   }
 
@@ -115,7 +114,7 @@ abstract class BuiltListMultimap<K, V> {
   ///
   /// Useful when producing or using APIs that need the [Map] interface. This
   /// differs from [toMap] where mutations are explicitly disallowed.
-  Map<K, Iterable<V>> asMap() => new Map<K, Iterable<V>>.unmodifiable(_map);
+  Map<K, Iterable<V>> asMap() => Map<K, Iterable<V>>.unmodifiable(_map);
 
   // ListMultimap.
 
@@ -132,7 +131,7 @@ abstract class BuiltListMultimap<K, V> {
   bool containsValue(Object value) => values.contains(value);
 
   /// As [ListMultimap.forEach].
-  void forEach(void f(K key, V value)) {
+  void forEach(void Function(K, V) f) {
     _map.forEach((key, values) {
       values.forEach((value) {
         f(key, value);
@@ -141,7 +140,7 @@ abstract class BuiltListMultimap<K, V> {
   }
 
   /// As [ListMultimap.forEachKey].
-  void forEachKey(void f(K key, Iterable<V> value)) {
+  void forEachKey(void Function(K, Iterable<V>) f) {
     _map.forEach((key, values) {
       f(key, values);
     });
@@ -156,9 +155,7 @@ abstract class BuiltListMultimap<K, V> {
   /// As [ListMultimap.keys], but result is stable; it always returns the same
   /// instance.
   Iterable<K> get keys {
-    if (_keys == null) {
-      _keys = _map.keys;
-    }
+    _keys ??= _map.keys;
     return _keys;
   }
 
@@ -168,9 +165,7 @@ abstract class BuiltListMultimap<K, V> {
   /// As [ListMultimap.values], but result is stable; it always returns the
   /// same instance.
   Iterable<V> get values {
-    if (_values == null) {
-      _values = _map.values.expand((x) => x);
-    }
+    _values ??= _map.values.expand((x) => x);
     return _values;
   }
 
@@ -178,11 +173,11 @@ abstract class BuiltListMultimap<K, V> {
 
   BuiltListMultimap._(this._map) {
     if (K == dynamic) {
-      throw new UnsupportedError('explicit key type required, '
+      throw UnsupportedError('explicit key type required, '
           'for example "new BuiltListMultimap<int, int>"');
     }
     if (V == dynamic) {
-      throw new UnsupportedError('explicit value type required,'
+      throw UnsupportedError('explicit value type required,'
           ' for example "new BuiltListMultimap<int, int>"');
     }
   }
@@ -193,12 +188,12 @@ class _BuiltListMultimap<K, V> extends BuiltListMultimap<K, V> {
   _BuiltListMultimap.withSafeMap(Map<K, BuiltList<V>> map) : super._(map);
 
   _BuiltListMultimap.copyAndCheck(Iterable keys, Function lookup)
-      : super._(new Map<K, BuiltList<V>>()) {
+      : super._(<K, BuiltList<V>>{}) {
     for (var key in keys) {
       if (key is K) {
-        _map[key] = new BuiltList<V>(lookup(key));
+        _map[key] = BuiltList<V>(lookup(key));
       } else {
-        throw new ArgumentError('map contained invalid key: $key');
+        throw ArgumentError('map contained invalid key: $key');
       }
     }
   }