+++ lib/src/map/built_map.dart
@@ -4,7 +4,7 @@
 
 part of built_collection.map;
 
-typedef Map<K, V> _MapFactory<K, V>();
+typedef _MapFactory<K, V> = Map<K, V> Function();
 
 /// The Built Collection [Map].
 ///
@@ -37,10 +37,9 @@ abstract class BuiltMap<K, V> {
     if (map is _BuiltMap && map.hasExactKeyAndValueTypes(K, V)) {
       return map as BuiltMap<K, V>;
     } else if (map is Map || map is BuiltMap) {
-      return new _BuiltMap<K, V>.copyAndCheckTypes(map.keys, (k) => map[k]);
+      return _BuiltMap<K, V>.copyAndCheckTypes(map.keys, (k) => map[k]);
     } else {
-      throw new ArgumentError(
-          'expected Map or BuiltMap, got ${map.runtimeType}');
+      throw ArgumentError('expected Map or BuiltMap, got ${map.runtimeType}');
     }
   }
 
@@ -54,40 +53,36 @@ abstract class BuiltMap<K, V> {
   ///
   /// Rejects nulls. Rejects keys and values of the wrong type.
   factory BuiltMap.from(Map map) {
-    return new _BuiltMap<K, V>.copyAndCheckTypes(map.keys, (k) => map[k]);
+    return _BuiltMap<K, V>.copyAndCheckTypes(map.keys, (k) => map[k]);
   }
 
   /// Instantiates with elements from a [Map<K, V>].
   ///
-  /// `K` and `V` must not be `dynamic`.
-  ///
-  /// Wrong: `new BuiltMap.of({1: '1', 2: '2', 3: '3'})`.
-  ///
-  /// Right: `new BuiltMap<int, String>.of({1: '1', 2: '2', 3: '3'})`.
+  /// `K` and `V` are inferred from `map`, and must not be `dynamic`.
   ///
   /// Rejects nulls. Rejects keys and values of the wrong type.
   factory BuiltMap.of(Map<K, V> map) {
-    return new _BuiltMap<K, V>.copyAndCheckForNull(map.keys, (k) => map[k]);
+    return _BuiltMap<K, V>.copyAndCheckForNull(map.keys, (k) => map[k]);
   }
 
   /// Creates a [MapBuilder], applies updates to it, and builds.
-  factory BuiltMap.build(updates(MapBuilder<K, V> builder)) =>
-      (new MapBuilder<K, V>()..update(updates)).build();
+  factory BuiltMap.build(Function(MapBuilder<K, V>) updates) =>
+      (MapBuilder<K, V>()..update(updates)).build();
 
   /// Converts to a [MapBuilder] for modification.
   ///
   /// The `BuiltMap` remains immutable and can continue to be used.
-  MapBuilder<K, V> toBuilder() => new MapBuilder<K, V>._fromBuiltMap(this);
+  MapBuilder<K, V> toBuilder() => MapBuilder<K, V>._fromBuiltMap(this);
 
   /// Converts to a [MapBuilder], applies updates to it, and builds.
-  BuiltMap<K, V> rebuild(updates(MapBuilder<K, V> builder)) =>
+  BuiltMap<K, V> rebuild(Function(MapBuilder<K, V>) updates) =>
       (toBuilder()..update(updates)).build();
 
   /// Returns as an immutable map.
   ///
   /// Useful when producing or using APIs that need the [Map] interface. This
   /// differs from [toMap] where mutations are explicitly disallowed.
-  Map<K, V> asMap() => new Map<K, V>.unmodifiable(_map);
+  Map<K, V> asMap() => Map<K, V>.unmodifiable(_map);
 
   /// Converts to a [Map].
   ///
@@ -97,7 +92,7 @@ abstract class BuiltMap<K, V> {
   ///
   /// This allows efficient use of APIs that ask for a mutable collection
   /// but don't actually mutate it.
-  Map<K, V> toMap() => new CopyOnWriteMap<K, V>(_map, _mapFactory);
+  Map<K, V> toMap() => CopyOnWriteMap<K, V>(_map, _mapFactory);
 
   /// Deep hashCode.
   ///
@@ -105,12 +100,10 @@ abstract class BuiltMap<K, V> {
   /// pairs in any order. Then, the `hashCode` is guaranteed to be the same.
   @override
   int get hashCode {
-    if (_hashCode == null) {
-      _hashCode = hashObjects(_map.keys
-          .map((key) => hash2(key.hashCode, _map[key].hashCode))
-          .toList(growable: false)
-            ..sort());
-    }
+    _hashCode ??= hashObjects(_map.keys
+        .map((key) => hash2(key.hashCode, _map[key].hashCode))
+        .toList(growable: false)
+          ..sort());
     return _hashCode;
   }
 
@@ -145,7 +138,7 @@ abstract class BuiltMap<K, V> {
   bool containsValue(Object value) => _map.containsValue(value);
 
   /// As [Map.forEach].
-  void forEach(void f(K key, V value)) {
+  void forEach(void Function(K, V) f) {
     _map.forEach(f);
   }
 
@@ -157,9 +150,7 @@ abstract class BuiltMap<K, V> {
 
   /// As [Map.keys], but result is stable; it always returns the same instance.
   Iterable<K> get keys {
-    if (_keys == null) {
-      _keys = _map.keys;
-    }
+    _keys ??= _map.keys;
     return _keys;
   }
 
@@ -169,9 +160,7 @@ abstract class BuiltMap<K, V> {
   /// As [Map.values], but result is stable; it always returns the same
   /// instance.
   Iterable<V> get values {
-    if (_values == null) {
-      _values = _map.values;
-    }
+    _values ??= _map.values;
     return _values;
   }
 
@@ -179,18 +168,18 @@ abstract class BuiltMap<K, V> {
   Iterable<MapEntry<K, V>> get entries => _map.entries;
 
   /// As [Map.map], but returns a [BuiltMap].
-  BuiltMap<K2, V2> map<K2, V2>(MapEntry<K2, V2> f(K key, V value)) =>
-      new _BuiltMap<K2, V2>.withSafeMap(null, _map.map(f));
+  BuiltMap<K2, V2> map<K2, V2>(MapEntry<K2, V2> Function(K, V) f) =>
+      _BuiltMap<K2, V2>.withSafeMap(null, _map.map(f));
 
   // Internal.
 
   BuiltMap._(this._mapFactory, this._map) {
     if (K == dynamic) {
-      throw new UnsupportedError(
+      throw UnsupportedError(
           'explicit key type required, for example "new BuiltMap<int, int>"');
     }
     if (V == dynamic) {
-      throw new UnsupportedError('explicit value type required,'
+      throw UnsupportedError('explicit value type required,'
           ' for example "new BuiltMap<int, int>"');
     }
   }
@@ -202,34 +191,40 @@ class _BuiltMap<K, V> extends BuiltMap<K, V> {
       : super._(mapFactory, map);
 
   _BuiltMap.copyAndCheckTypes(Iterable keys, Function lookup)
-      : super._(null, new Map<K, V>()) {
+      : super._(null, <K, V>{}) {
     for (var key in keys) {
       if (key is K) {
         var value = lookup(key);
         if (value is V) {
           _map[key] = value;
         } else {
-          throw new ArgumentError('map contained invalid value: $value');
+          throw ArgumentError('map contained invalid value: $value');
         }
       } else {
-        throw new ArgumentError('map contained invalid key: $key');
+        throw ArgumentError('map contained invalid key: $key');
       }
     }
   }
 
-  _BuiltMap.copyAndCheckForNull(Iterable<K> keys, V lookup(K value))
-      : super._(null, new Map<K, V>()) {
+  _BuiltMap.copyAndCheckForNull(Iterable<K> keys, V Function(K) lookup)
+      : super._(null, <K, V>{}) {
     for (var key in keys) {
       if (identical(key, null)) {
-        throw new ArgumentError('map contained invalid key: null');
+        throw ArgumentError('map contained invalid key: null');
       }
       var value = lookup(key);
       if (value == null) {
-        throw new ArgumentError('map contained invalid value: null');
+        throw ArgumentError('map contained invalid value: null');
       }
       _map[key] = value;
     }
   }
 
   bool hasExactKeyAndValueTypes(Type key, Type value) => K == key && V == value;
+}
+
+/// Extensions for [BuiltMap] on [Map].
+extension BuiltMapExtension<K, V> on Map<K, V> {
+  /// Converts to a [BuiltMap].
+  BuiltMap<K, V> build() => BuiltMap<K, V>.of(this);
 }