+++ lib/src/set/built_set.dart
@@ -4,7 +4,7 @@
 
 part of built_collection.set;
 
-typedef Set<E> _SetFactory<E>();
+typedef _SetFactory<E> = Set<E> Function();
 
 /// The Built Collection [Set].
 ///
@@ -29,8 +29,7 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
   /// Right: `new BuiltSet<int>([1, 2, 3])`.
   ///
   /// Rejects nulls. Rejects elements of the wrong type.
-  factory BuiltSet([Iterable iterable = const []]) =>
-      new BuiltSet.from(iterable);
+  factory BuiltSet([Iterable iterable = const []]) => BuiltSet.from(iterable);
 
   /// Instantiates with elements from an [Iterable].
   ///
@@ -45,7 +44,7 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
     if (iterable is _BuiltSet && iterable.hasExactElementType(E)) {
       return iterable as BuiltSet<E>;
     } else {
-      return new _BuiltSet<E>.copyAndCheckTypes(iterable);
+      return _BuiltSet<E>.copyAndCheckTypes(iterable);
     }
   }
 
@@ -58,25 +57,25 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
     if (iterable is _BuiltSet<E> && iterable.hasExactElementType(E)) {
       return iterable;
     } else {
-      return new _BuiltSet<E>.copyAndCheckForNull(iterable);
+      return _BuiltSet<E>.copyAndCheckForNull(iterable);
     }
   }
 
   /// Creates a [SetBuilder], applies updates to it, and builds.
-  factory BuiltSet.build(updates(SetBuilder<E> builder)) =>
-      (new SetBuilder<E>()..update(updates)).build();
+  factory BuiltSet.build(Function(SetBuilder<E>) updates) =>
+      (SetBuilder<E>()..update(updates)).build();
 
   /// Converts to a [SetBuilder] for modification.
   ///
   /// The `BuiltSet` remains immutable and can continue to be used.
-  SetBuilder<E> toBuilder() => new SetBuilder<E>._fromBuiltSet(this);
+  SetBuilder<E> toBuilder() => SetBuilder<E>._fromBuiltSet(this);
 
   /// Converts to a [SetBuilder], applies updates to it, and builds.
-  BuiltSet<E> rebuild(updates(SetBuilder<E> builder)) =>
+  BuiltSet<E> rebuild(Function(SetBuilder<E>) updates) =>
       (toBuilder()..update(updates)).build();
 
   @override
-  BuiltList<E> toBuiltList() => new BuiltList<E>(this);
+  BuiltList<E> toBuiltList() => BuiltList<E>(this);
 
   @override
   BuiltSet<E> toBuiltSet() => this;
@@ -87,10 +86,8 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
   /// any order. Then, the `hashCode` is guaranteed to be the same.
   @override
   int get hashCode {
-    if (_hashCode == null) {
-      _hashCode = hashObjects(
-          _set.map((e) => e.hashCode).toList(growable: false)..sort());
-    }
+    _hashCode ??= hashObjects(
+        _set.map((e) => e.hashCode).toList(growable: false)..sort());
     return _hashCode;
   }
 
@@ -114,7 +111,7 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
   ///
   /// Useful when producing or using APIs that need the [Set] interface. This
   /// differs from [toSet] where mutations are explicitly disallowed.
-  Set<E> asSet() => new UnmodifiableSetView<E>(_set);
+  Set<E> asSet() => UnmodifiableSetView<E>(_set);
 
   // Set.
 
@@ -127,18 +124,18 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
 
   /// As [Set.difference] but takes and returns a `BuiltSet<E>`.
   BuiltSet<E> difference(BuiltSet<Object> other) =>
-      new _BuiltSet<E>.withSafeSet(_setFactory, _set.difference(other._set));
+      _BuiltSet<E>.withSafeSet(_setFactory, _set.difference(other._set));
 
   /// As [Set.intersection] but takes and returns a `BuiltSet<E>`.
   BuiltSet<E> intersection(BuiltSet<Object> other) =>
-      new _BuiltSet<E>.withSafeSet(_setFactory, _set.intersection(other._set));
+      _BuiltSet<E>.withSafeSet(_setFactory, _set.intersection(other._set));
 
   /// As [Set.lookup].
   E lookup(Object object) => _set.lookup(object);
 
   /// As [Set.union] but takes and returns a `BuiltSet<E>`.
   BuiltSet<E> union(BuiltSet<E> other) =>
-      new _BuiltSet<E>.withSafeSet(_setFactory, _set.union(other._set));
+      _BuiltSet<E>.withSafeSet(_setFactory, _set.union(other._set));
 
   // Iterable.
 
@@ -155,35 +152,35 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
   Iterable<T> whereType<T>() => _set.whereType<T>();
 
   @override
-  Iterable<T> map<T>(T f(E e)) => _set.map(f);
+  Iterable<T> map<T>(T Function(E) f) => _set.map(f);
 
   @override
-  Iterable<E> where(bool test(E element)) => _set.where(test);
+  Iterable<E> where(bool Function(E) test) => _set.where(test);
 
   @override
-  Iterable<T> expand<T>(Iterable<T> f(E e)) => _set.expand(f);
+  Iterable<T> expand<T>(Iterable<T> Function(E) f) => _set.expand(f);
 
   @override
   bool contains(Object element) => _set.contains(element);
 
   @override
-  void forEach(void f(E element)) => _set.forEach(f);
+  void forEach(void Function(E) f) => _set.forEach(f);
 
   @override
-  E reduce(E combine(E value, E element)) => _set.reduce(combine);
+  E reduce(E Function(E, E) combine) => _set.reduce(combine);
 
   @override
-  T fold<T>(T initialValue, T combine(T previousValue, E element)) =>
+  T fold<T>(T initialValue, T Function(T, E) combine) =>
       _set.fold(initialValue, combine);
 
   @override
-  bool every(bool test(E element)) => _set.every(test);
+  bool every(bool Function(E) test) => _set.every(test);
 
   @override
   String join([String separator = '']) => _set.join(separator);
 
   @override
-  bool any(bool test(E element)) => _set.any(test);
+  bool any(bool Function(E) test) => _set.any(test);
 
   /// As [Iterable.toSet].
   ///
@@ -194,7 +191,7 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
   /// This allows efficient use of APIs that ask for a mutable collection
   /// but don't actually mutate it.
   @override
-  Set<E> toSet() => new CopyOnWriteSet<E>(_set, _setFactory);
+  Set<E> toSet() => CopyOnWriteSet<E>(_set, _setFactory);
 
   @override
   List<E> toList({bool growable = true}) => _set.toList(growable: growable);
@@ -209,13 +206,13 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
   Iterable<E> take(int n) => _set.take(n);
 
   @override
-  Iterable<E> takeWhile(bool test(E value)) => _set.takeWhile(test);
+  Iterable<E> takeWhile(bool Function(E) test) => _set.takeWhile(test);
 
   @override
   Iterable<E> skip(int n) => _set.skip(n);
 
   @override
-  Iterable<E> skipWhile(bool test(E value)) => _set.skipWhile(test);
+  Iterable<E> skipWhile(bool Function(E) test) => _set.skipWhile(test);
 
   @override
   E get first => _set.first;
@@ -227,15 +224,15 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
   E get single => _set.single;
 
   @override
-  E firstWhere(bool test(E element), {E orElse()}) =>
+  E firstWhere(bool Function(E) test, {E Function() orElse}) =>
       _set.firstWhere(test, orElse: orElse);
 
   @override
-  E lastWhere(bool test(E element), {E orElse()}) =>
+  E lastWhere(bool Function(E) test, {E Function() orElse}) =>
       _set.lastWhere(test, orElse: orElse);
 
   @override
-  E singleWhere(bool test(E element), {E orElse()}) =>
+  E singleWhere(bool Function(E) test, {E Function() orElse}) =>
       _set.singleWhere(test, orElse: orElse);
 
   @override
@@ -245,7 +242,7 @@ abstract class BuiltSet<E> implements Iterable<E>, BuiltIterable<E> {
 
   BuiltSet._(this._setFactory, this._set) {
     if (E == dynamic) {
-      throw new UnsupportedError(
+      throw UnsupportedError(
           'explicit element type required, for example "new BuiltSet<int>"');
     }
   }
@@ -256,21 +253,20 @@ class _BuiltSet<E> extends BuiltSet<E> {
   _BuiltSet.withSafeSet(_SetFactory<E> setFactory, Set<E> set)
       : super._(setFactory, set);
 
-  _BuiltSet.copyAndCheckTypes(Iterable iterable) : super._(null, new Set<E>()) {
+  _BuiltSet.copyAndCheckTypes(Iterable iterable) : super._(null, <E>{}) {
     for (var element in iterable) {
       if (element is E) {
         _set.add(element);
       } else {
-        throw new ArgumentError('iterable contained invalid element: $element');
+        throw ArgumentError('iterable contained invalid element: $element');
       }
     }
   }
 
-  _BuiltSet.copyAndCheckForNull(Iterable iterable)
-      : super._(null, new Set<E>()) {
+  _BuiltSet.copyAndCheckForNull(Iterable iterable) : super._(null, <E>{}) {
     for (var element in iterable) {
       if (identical(element, null)) {
-        throw new ArgumentError('iterable contained invalid element: null');
+        throw ArgumentError('iterable contained invalid element: null');
       } else {
         _set.add(element);
       }
@@ -278,4 +274,21 @@ class _BuiltSet<E> extends BuiltSet<E> {
   }
 
   bool hasExactElementType(Type type) => E == type;
+}
+
+/// Extensions for [BuiltSet] on [Set].
+extension BuiltSetExtension<T> on Set<T> {
+  /// Converts to a [BuiltSet].
+  BuiltSet<T> build() {
+    // We know a `Set` is not a `BuiltSet`, so we have to copy.
+    return _BuiltSet<T>.copyAndCheckForNull(this);
+  }
+}
+
+/// Extensions for [BuiltSet] on [Iterable].
+extension BuiltSetIterableExtension<E> on Iterable<E> {
+  /// Converts to a [BuiltSet].
+  ///
+  /// Just returns the [Iterable] if it is already a `BuiltSet<E>`.
+  BuiltSet<E> toBuiltSet() => BuiltSet<E>.of(this);
 }